jobs/requirements.txt
google-api-python-client
requests
httplib2
elasticsearch
python-dotenv
